+++ lib/Jira.js
